Use UUIDs instead of IDs (#233)
* Implement migrations for uuid

* Add serialize functions, change repos and routers to use uuid, and update tests

* Change documentation

* Edit test

+++ src/repos/GroupsRepo.js
@@ -38,7 +38,7 @@ const createGroup = async (
       throw LogUtils.logErr(`Group code is already in use: ${code}`);
     }
     await db().save(group);
-    groupCodes[group.code] = group.id;
+    groupCodes[group.code] = group.uuid;
     return group;
   } catch (e) {
     throw LogUtils.logErr('Problem creating group', e, {
@@ -61,63 +61,69 @@ const createCode = (): string => {
 };
 
 /**
- * Get a group by id
+ * Get a group by UUID
  * @function
- * @param {number} id - ID of group to fetch
- * @return {?Group} Group with specified id
+ * @param {string} id - UUID of group to fetch
+ * @return {?Group} Group with specified UUID
  */
-const getGroupByID = async (id: number): Promise<?Group> => {
+const getGroupByID = async (id: string): Promise<?Group> => {
   try {
-    return await db().findOne(id);
+    return await db().createQueryBuilder('groups')
+      .where('groups.uuid = :groupID')
+      .setParameters({ groupID: id })
+      .getOne();
   } catch (e) {
-    throw LogUtils.logErr(`Problem getting group by id: ${id}`, e);
+    throw LogUtils.logErr(`Problem getting group by UUID: ${id}`, e);
   }
 };
 
 /**
- * Get a group id by the group's unique code
+ * Get a group UUID by the group's unique code
  * @function
  * @param {string} code - Unique code of group to fetch
- * @return {?number} ID of group with given code
+ * @return {?string} UUID of group with given code
  */
 const getGroupID = async (code: string) => {
   const group = await db().createQueryBuilder('groups')
     .where('groups.code = :groupCode')
     .setParameters({ groupCode: code })
     .getOne();
-  return group ? group.id : null;
+  return group ? group.uuid : null;
 };
 
 /**
  * Delete a group
  * @function
- * @param {number} id - ID of group to delete
+ * @param {string} id - UUID of group to delete
  */
-const deleteGroupByID = async (id: number) => {
+const deleteGroupByID = async (id: string) => {
   try {
-    const group = await db().findOne(id);
-    delete groupCodes[group.code];
-    await db().remove(group);
+    const group = await getGroupByID(id);
+    
+    if (group) {
+      delete groupCodes[group.code];
+      await db().remove(group);
+    }
   } catch (e) {
-    throw LogUtils.logErr(`Problem deleting group by id: ${id}`, e);
+    throw LogUtils.logErr(`Problem deleting group by UUID: ${id}`, e);
   }
 };
 
 /**
- * Update a group by group ID
+ * Update a group by group UUID
  * @function
- * @param {number} id - ID of group to update
+ * @param {string} id - UUID of group to update
  * @param {?name} name - New group name
  * @return {?Group} Updated group
  */
-const updateGroupByID = async (id: number, name: ?string):
+const updateGroupByID = async (id: string, name: ?string):
   Promise<?Group> => {
   try {
     const group = await db().createQueryBuilder('groups')
       .leftJoinAndSelect('groups.admins', 'admins')
       .leftJoinAndSelect('groups.members', 'members')
       .leftJoinAndSelect('groups.polls', 'polls')
-      .where('groups.id = :groupID', { groupID: id })
+      .where('groups.uuid = :groupID', { groupID: id })
       .getOne();
 
     if (name) group.name = name;
@@ -131,22 +137,19 @@ const updateGroupByID = async (id: number, name: ?string):
 /**
  * Add users (admins or members) to a group
  * @function
- * @param {number} id - ID of group to add users
- * @param {string[]} googleIDs - List of user's google ids to add
+ * @param {string} id - UUID of group to add users
+ * @param {string[]} googleIDs - List of user's googleIDs to add
  * @param {string} [role] - Specifies whether to add the users as members or admins
  * @return {?Group} Group that users were added to
  */
-const addUsersByGoogleIDs = async (
-  id: number,
-  googleIDs: string[],
-  role: ?string,
-): Promise<?Group> => {
+const addUsersByGoogleIDs = async (id: string, googleIDs: string[],
+  role: ?string): Promise<?Group> => {
   try {
     const group = await db().createQueryBuilder('groups')
       .leftJoinAndSelect('groups.admins', 'admins')
       .leftJoinAndSelect('groups.members', 'members')
       .leftJoinAndSelect('groups.polls', 'polls')
-      .where('groups.id = :groupID')
+      .where('groups.uuid = :groupID')
       .setParameters({ groupID: id })
       .getOne();
     if (group) {
@@ -176,31 +179,28 @@ const addUsersByGoogleIDs = async (
 /**
  * Add users (admins or members) to a group
  * @function
- * @param {number} id - ID of group to add users
- * @param {number[]} userIDs - List of user ids to add
+ * @param {string} id - UUID of group to add users
+ * @param {string[]} userIDs - List of user UUIDs to add
  * @param {string} [role] - Specifies whether to add the users as members or admins
  * @return {?Group} Group that users were added to
  */
-const addUsersByIDs = async (
-  id: number,
-  userIDs: number[],
-  role: ?string,
-): Promise<?Group> => {
+const addUsersByIDs = async (id: string, userIDs: string[],
+  role: ?string): Promise<?Group> => {
   try {
     const group = await db().createQueryBuilder('groups')
       .leftJoinAndSelect('groups.admins', 'admins')
       .leftJoinAndSelect('groups.members', 'members')
       .leftJoinAndSelect('groups.polls', 'polls')
-      .where('groups.id = :groupID')
+      .where('groups.uuid = :groupID')
       .setParameters({ groupID: id })
       .getOne();
     if (group) {
       if (role === constants.USER_TYPES.ADMIN) {
-        const currAdminIDs = group.admins.map(admin => admin.id);
+        const currAdminIDs = group.admins.map(admin => admin.uuid);
         const admins = await UsersRepo.getUsersFromIDs(userIDs, currAdminIDs);
         group.admins = group.admins.concat(admins);
       } else {
-        const currMemberIDs = group.members.map(member => member.id);
+        const currMemberIDs = group.members.map(member => member.uuid);
         const members = await UsersRepo.getUsersFromIDs(userIDs, currMemberIDs);
         group.members = group.members.concat(members);
       }
@@ -208,57 +208,57 @@ const addUsersByIDs = async (
     await db().save(group);
     return group;
   } catch (e) {
-    throw LogUtils.logErr(`Problem adding users to group ${id} by ids`, e, { userIDs, role });
+    console.log(e);
+    throw LogUtils.logErr(`Problem adding users to group ${id} by UUIDs`, e, { userIDs, role });
   }
 };
 
 /**
  * Remove user from group
  * @function
- * @param {number} id - ID of group to remove user from
- * @param {User} user - User to remove from group
+ * @param {string} id - UUID of group to remove user from
+ * @param {string[]} userIDs - List of user UUIDs to remove from group
  * @param {string} [role] - Role to remove user from
  * @return {?Group} Group without specified user
  */
-const removeUserByGroupID = async (
-  id: number,
-  user: User,
-  role: ?string,
-): Promise<?Group> => {
+const removeUserByGroupID = async (id: string, userIDs: string[], role: ?string):
+  Promise<?Group> => {
   try {
     const group = await db().createQueryBuilder('groups')
       .leftJoinAndSelect('groups.admins', 'admins')
       .leftJoinAndSelect('groups.members', 'members')
       .leftJoinAndSelect('groups.polls', 'polls')
-      .where('groups.id = :groupID')
+      .where('groups.uuid = :groupID')
       .setParameters({ groupID: id })
       .getOne();
-    if (user) {
+
+    if (group) {
       if (role === constants.USER_TYPES.ADMIN) {
-        group.admins = group.admins.filter(admin => admin.googleID !== user.googleID);
+        group.admins = group.admins.filter(admin => !userIDs.includes(admin.uuid));
       } else {
-        group.members = group.members.filter(member => member.googleID !== user.googleID);
+        group.members = group.members.filter(member => !userIDs.includes(member.uuid));
       }
       await db().save(group);
     }
     return group;
   } catch (e) {
-    throw LogUtils.logErr(`Problem removing user from group by id: ${id}`, e, { user, role });
+    throw LogUtils.logErr(`Problem removing user from group by UUID: ${id}`, e, { userIDs, role });
   }
 };
 
 /**
  * Checks if user is an admin of given group
  * @function
- * @param {number} id - ID of group to check admins
+ * @param {string} id - UUID of group to check admins
  * @param {User} user - User that we want to check if they're an admin
  * @return {?boolean} Whether the given user is an admin of the given group
  */
-const isAdmin = async (id: number, user: User): Promise<?boolean> => {
+const isAdmin = async (id: string, user: User):
+  Promise<?boolean> => {
   try {
     const group = await db().createQueryBuilder('groups')
       .leftJoinAndSelect('groups.admins', 'admins')
-      .where('groups.id = :groupID')
+      .where('groups.uuid = :groupID')
       .setParameters({ groupID: id })
       .getOne();
     const admin = group.admins.find(x => x.googleID === user.googleID);
@@ -271,15 +271,16 @@ const isAdmin = async (id: number, user: User): Promise<?boolean> => {
 /**
  * Checks if user is a member of given group
  * @function
- * @param {number} id - ID of group to check members
+ * @param {string} id - UUID of group to check members
  * @param {User} user - User that we want to check if they're a member
  * @return {?boolean} Whether the given user is a member of the given group
  */
-const isMember = async (id: number, user: User): Promise<?boolean> => {
+const isMember = async (id: string, user: User):
+  Promise<?boolean> => {
   try {
     const group = await db().createQueryBuilder('groups')
       .leftJoinAndSelect('groups.members', 'members')
-      .where('groups.id = :groupID')
+      .where('groups.uuid = :groupID')
       .setParameters({ groupID: id })
       .getOne();
     const member = group.members.find(x => x.googleID === user.googleID);
@@ -292,16 +293,17 @@ const isMember = async (id: number, user: User): Promise<?boolean> => {
 /**
  * Get users from a group
  * @function
- * @param {number} id - ID of group to get users
+ * @param {string} id - UUID of group to get users
  * @param {string} [role] - Specifies if we only want users of a certain role
  * @return {User[]} List of specified user from group
  */
-const getUsersByGroupID = async (id: number, role: ?string): Promise<Array<?User>> => {
+const getUsersByGroupID = async (id: string, role: ?string):
+  Promise<Array<?User>> => {
   try {
     const group = await db().createQueryBuilder('groups')
       .leftJoinAndSelect('groups.admins', 'admins')
       .leftJoinAndSelect('groups.members', 'members')
-      .where('groups.id = :groupID')
+      .where('groups.uuid = :groupID')
       .setParameters({ groupID: id })
       .getOne();
     if (role === constants.USER_TYPES.ADMIN) {
@@ -312,7 +314,8 @@ const getUsersByGroupID = async (id: number, role: ?string): Promise<Array<?User
     }
     return group.admins.concat(group.members);
   } catch (e) {
-    throw LogUtils.logErr(`Problem getting users for group by id: ${id}`, e, { role });
+    console.log(e);
+    throw LogUtils.logErr(`Problem getting users for group by UUID: ${id}`, e, { role });
   }
 };
 
@@ -320,18 +323,16 @@ const getUsersByGroupID = async (id: number, role: ?string): Promise<Array<?User
  * Gets polls from a group sorted by creation date in ascending order.
  * By default will hide poll results if poll is not shared.
  * @function
- * @param {number} id - ID of group to fetch polls from
+ * @param {string} id - UUID of group to fetch polls from
  * @param {boolean} hideUnsharedResults - Whether to return unaltered poll results for admins
  * @return {Poll[]} List of polls from group
  */
-const getPolls = async (
-  id: number,
-  hideUnsharedResults: ?boolean = true,
-): Promise<Array<?Poll>> => {
+const getPolls = async (id: string, hideUnsharedResults: ?boolean = true):
+  Promise<Array<?Poll>> => {
   try {
     const group = await db().createQueryBuilder('groups')
       .leftJoinAndSelect('groups.polls', 'polls')
-      .where('groups.id = :groupID')
+      .where('groups.uuid = :groupID')
       .setParameters({ groupID: id })
       .orderBy('polls.createdAt', 'ASC')
       .getOne();
@@ -356,12 +357,12 @@ const getPolls = async (
 /**
  * Get time of latest activity of a group
  * @function
- * @param {number} id - ID of group to get latest activity
+ * @param {string} id - UUID of group to get latest activity
  * @return {string} Time stamp of when the group was last updated
  */
-const latestActivityByGroupID = async (id: number): Promise<string> => {
+const latestActivityByGroupID = async (id: string): Promise<string> => {
   try {
-    const group = await db().findOne(id);
+    const group = await getGroupByID(id);
     if (!group) throw LogUtils.logErr(`Can't find group by id: ${id}`);
     return await getPolls(id).then((polls: Array<?Poll>) => {
       const latestPoll = polls.slice(-1).pop();
@@ -373,7 +374,7 @@ const latestActivityByGroupID = async (id: number): Promise<string> => {
         : latestPoll.updatedAt;
     });
   } catch (e) {
-    throw LogUtils.logErr(`Problem getting latest activity from group by id: ${id}`, e);
+    throw LogUtils.logErr(`Problem getting latest activity from group by UUID: ${id}`, e);
   }
 };
