+++ src/repos/UserSessionsRepo.js
@@ -22,8 +22,8 @@ const createOrUpdateSession = async (
   refreshToken: ?string,
 ): Promise<UserSession> => {
   const optionalSession = await db().createQueryBuilder('usersessions')
-    .where('usersessions.user = :userID', { userID: user.id })
-    .innerJoinAndSelect('usersessions.user', 'users')
+    .innerJoin('usersessions.user', 'user', 'user.uuid = :userID')
+    .setParameters({ userID: user.uuid })
     .getOne();
   return db().save(
     optionalSession
@@ -86,29 +86,32 @@ const verifySession = async (accessToken: string): Promise<boolean> => {
 /**
  * Delete a session
  * @function
- * @param {number} id - ID of session to delete
+ * @param {string} id - UUID of session to delete
  */
-const deleteSession = async (id: number) => {
+const deleteSession = async (id: string) => {
   try {
-    const session = await db().findOne(id);
-    await db().remove(session);
+    const session = await db().createQueryBuilder('usersessions')
+      .where('usersessions.uuid = :sessionID')
+      .setParameters({ sessionID: id })
+      .getOne();
+    if (session) await db().remove(session);
   } catch (e) {
-    throw LogUtils.logErr(`Problem deleting session by id: ${id}`, e);
+    throw LogUtils.logErr(`Problem deleting session by UUID: ${id}`, e);
   }
 };
 
 /**
  * Delete the session for a user
  * @function
- * @param {number} userID - ID of use to delete the session for
+ * @param {string} userID - UUID of user to delete the session for
  */
-const deleteSessionFromUserID = async (userID: number) => {
+const deleteSessionFromUserID = async (userID: string) => {
   try {
     const session = await db().createQueryBuilder('usersessions')
-      .innerJoin('usersessions.user', 'user', 'user.id = :userID')
+      .innerJoin('usersessions.user', 'user', 'user.uuid = :userID')
       .setParameters({ userID })
       .getOne();
-    if (session) db().remove(session);
+    if (session) await db().remove(session);
   } catch (e) {
     throw LogUtils.logErr(`Problem deleting session by user: ${userID}`, e);
   }
@@ -132,12 +135,7 @@ const createUserAndInitializeSession = async (login: LoginTicket): Promise<Objec
     user = await UsersRepo.createUserWithFields(googleID, first, last, email);
   }
   const session = await createOrUpdateSession(user, null, null);
-  return {
-    accessToken: session.sessionToken,
-    refreshToken: session.updateToken,
-    sessionExpiration: session.expiresAt,
-    isActive: session.isActive,
-  };
+  return session.serialize();
 };
 
 export default {
